svg_to_video/requirements.txt
fastapi
uvicorn
requests
cairosvg
pillow
python-multipart
aiohttp
supabase
python-dotenv